Add "'cobalt_code': 1" to dom.gyp, fix Windows build
To make cobalt_code 1 work, the following changes are made:
- Added MSVC_PUSH_DISABLE_WARNING 4355 to anywhere "this" is used in base
  member initializer list.
- Disabled warning C4702: unreachable code in xtree.
- Added static_cast to all places where implicit conversion will cause warnings,
  or change variable type if it makes sense.
- Fixed all unreferenced parameter warnings.

b/22613037

+++ cobalt/dom/event.cc
@@ -24,18 +24,21 @@ namespace cobalt {
 namespace dom {
 
 Event::Event(UninitializedFlag uninitialized_flag)
-    : event_phase_(kNone), time_stamp_(base::Time::Now().ToJsTime()) {
+    : event_phase_(kNone),
+      time_stamp_(static_cast<uint64>(base::Time::Now().ToJsTime())) {
   UNREFERENCED_PARAMETER(uninitialized_flag);
   InitEvent("", false, false);
 }
 
 Event::Event(const std::string& type)
-    : event_phase_(kNone), time_stamp_(base::Time::Now().ToJsTime()) {
+    : event_phase_(kNone),
+      time_stamp_(static_cast<uint64>(base::Time::Now().ToJsTime())) {
   InitEvent(type, false, false);
 }
 
 Event::Event(const std::string& type, Bubbles bubbles, Cancelable cancelable)
-    : event_phase_(kNone), time_stamp_(base::Time::Now().ToJsTime()) {
+    : event_phase_(kNone),
+      time_stamp_(static_cast<uint64>(base::Time::Now().ToJsTime())) {
   InitEvent(type, bubbles == kBubbles, cancelable == kCancelable);
 }
